requirements.txt
attrs==23.2.0
branca==0.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
coverage==7.5.3
cycler==0.12.1
exceptiongroup==1.2.1
folium==0.17.0
fonttools==4.53.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
xyzservices==2024.6.0
